Created a ChatService and moved relevant elements from the SessionService
