+++ client/src/main/java/org/schlunzis/kurtama/client/service/IChatService.java
@@ -1,0 +1,13 @@
+package org.schlunzis.kurtama.client.service;
+
+import javafx.collections.ObservableList;
+
+public interface IChatService {
+
+    ObservableList<String> getChatMessages();
+
+    String getCurrentUsername();
+
+    void sendMessage(String nickname, String message);
+
+}
