+++ client/src/main/java/org/schlunzis/kurtama/client/service/impl/SessionService.java
@@ -10,7 +10,6 @@ import org.schlunzis.kurtama.common.IUser;
 import org.schlunzis.kurtama.common.LobbyInfo;
 import org.schlunzis.kurtama.common.messages.authentication.login.LoginSuccessfulResponse;
 import org.schlunzis.kurtama.common.messages.authentication.logout.LogoutSuccessfulResponse;
-import org.schlunzis.kurtama.common.messages.chat.ServerChatMessage;
 import org.schlunzis.kurtama.common.messages.lobby.server.JoinLobbySuccessfullyResponse;
 import org.schlunzis.kurtama.common.messages.lobby.server.LeaveLobbySuccessfullyResponse;
 import org.schlunzis.kurtama.common.messages.lobby.server.LobbyCreatedSuccessfullyResponse;
@@ -20,7 +19,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.Optional;
-import java.util.UUID;
 
 /**
  * A component to save information about the current session. Like the user the client is logged in as.
@@ -29,11 +27,8 @@ import java.util.UUID;
 @Service
 public class SessionService implements ISessionService {
 
-    private static final UUID GLOBAL_CHAT_ID = new UUID(0, 0);
 
     private final ObservableList<LobbyInfo> lobbyList = FXCollections.observableList(new ArrayList<>());
-    private final ObservableList<String> chatMessages = FXCollections.observableList(new ArrayList<>());
-    private UUID currentChatID = GLOBAL_CHAT_ID;
     private Optional<IUser> currentUser = Optional.empty();
     private Optional<ILobby> currentLobby = Optional.empty();
 
@@ -46,25 +41,21 @@ public class SessionService implements ISessionService {
     @EventListener
     public void onLogoutSuccessfulResponse(LogoutSuccessfulResponse lsr) {
         currentUser = Optional.empty();
-        currentChatID = GLOBAL_CHAT_ID;
     }
 
     @EventListener
     public void onLobbyCreatedSuccessfullyResponse(LobbyCreatedSuccessfullyResponse lcsr) {
         currentLobby = Optional.of(lcsr.lobby());
-        currentChatID = lcsr.lobby().getChatID();
     }
 
     @EventListener
     public void onJoinLobbySuccessfullyResponse(JoinLobbySuccessfullyResponse jlsr) {
         currentLobby = Optional.of(jlsr.lobby());
-        currentChatID = jlsr.lobby().getChatID();
     }
 
     @EventListener
     public void onLeaveLobbySuccessfullyResponse(LeaveLobbySuccessfullyResponse llsr) {
         currentLobby = Optional.empty();
-        currentChatID = GLOBAL_CHAT_ID;
     }
 
     @EventListener
@@ -72,13 +63,6 @@ public class SessionService implements ISessionService {
         Platform.runLater(() ->
                 lobbyList.setAll(llim.lobbies())
         );
-    }
-
-    @EventListener
-    public void onServerChatMessage(ServerChatMessage message) {
-        if (message.getChatID().equals(currentChatID)) {
-            Platform.runLater(() -> chatMessages.add(message.getNickname() + ": " + message.getMessage()));
-        }
     }
 
 }
